# swagger.yaml
openapi: "3.0.0"
paths:
...
  /comment/create:
    post:
      operationId: "CommentController_createComment"
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCommentDto"
      responses:
        201:
          description: "Comentario creado con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateCommentResponseDto"
        404:
          description: "Perfil no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /comment/get-by-profile/{id}:
    get:
      operationId: "CommentController_getCommentsByProfileId"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Comentarios obtenidos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetByProfileCommentResponseDto"
        404:
          description: "Perfil no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /comment/toggle-solve/{id}:
    patch:
      operationId: "CommentController_toggleSolveComment"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Cambio de estado en la resolución del comentario"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ToggleSolveCommentResponseDto"
        404:
          description: "Comentario no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        409:
          description: "El comentario no es resoluble"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
...
  /event-folder/create:
    post:
      operationId: "EventFolderController_createEventFolder"
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventFolderDto"
      responses:
        201:
          description: "Carpeta de evento creada con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateEventFolderResponseDto"
        409:
          description: "Error en la creación de la carpeta de evento"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
...
  /event-folder/{id}:
    get:
      operationId: "EventFolderController_getEventFolderById"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Carpeta de evento obtenida con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetByIdEventFolderResponseDto"
        404:
          description: "Carpeta de evento no encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /event-folder/update/{id}:
    patch:
      operationId: "EventFolderController_updateEventFolder"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventFolderDto"
      responses:
        200:
          description: "Carpeta de evento actualizada con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateEventFolderResponseDto"
        404:
          description: "Carpeta de evento no encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /event-folder/delete/{id}:
    delete:
      operationId: "EventFolderController_deleteEventFolder"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Carpeta de evento eliminada con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteEventFolderResponseDto"
        404:
          description: "Carpeta de evento no encontrada"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /event/create:
    post:
      operationId: "EventController_create"
      parameters: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateEventDto"
      responses:
        201:
          description: "Evento creado con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateEventResponseDto"
        409:
          description: "Error en la creación del evento"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /event/all:
    get:
      operationId: "EventController_findAll"
      parameters: []
      responses:
        200:
          description: "Eventos obtenidos con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetAllEventsResponseDto"
        404:
          description: "No se encontraron eventos"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
  /event/{id}:
    get:
      operationId: "EventController_findById"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        200:
          description: "Evento obtenido con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GetByIdEventResponseDto"
        404:
          description: "Evento no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    patch:
      operationId: "EventController_update"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateEventDto"
      responses:
        200:
          description: "Evento actualizado con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UpdateEventResponseDto"
        404:
          description: "Evento no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
    delete:
      operationId: "EventController_remove"
      parameters:
        - name: "id"
          required: true
          in: "path"
          schema:
            type: "string"
      responses:
        404:
          description: "Evento no encontrado"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorDto"
        410:
          description: "Evento eliminado con éxito"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteEventResponseDto"
...
info:
  title: "Expo Backend"
  description: "Backend de las aplicaciones de Expo"
  version: "1.0"
  contact: {}
...
servers: []
components:
  schemas:
...
    ErrorDto:
      type:
        - "object"
      properties:
        message:
          type:
            - "array"
          items:
            type:
              - "string"
        statusCode:
          type:
            - "number"
        error:
          type:
            - "string"
      required:
        - "message"
        - "statusCode"
        - "error"
...
    CreateCommentDto:
      type:
        - "object"
      properties:
        content:
          type:
            - "string"
          minLength: 1
        profileId:
          type:
            - "string"
          format: "uuid"
        isSolvable:
          default: false
          type:
            - "boolean"
      required:
        - "content"
        - "profileId"
    CreateCommentResponseDto:
      type:
        - "object"
      properties:
        id:
          type:
            - "string"
          format: "uuid"
        content:
          type:
            - "string"
          minLength: 1
        createdBy:
          type:
            - "string"
          format: "uuid"
        profileId:
          type:
            - "string"
          format: "uuid"
        isSolvable:
          default: false
          type:
            - "boolean"
        isSolved:
          default: false
          type:
            - "boolean"
        solvedAt:
          type:
            - "string"
            - "null"
          format: "date-time"
        solvedBy:
          type:
            - "string"
          format: "uuid"
        created_at:
          type:
            - "string"
          format: "date-time"
        updated_at:
          type:
            - "string"
          format: "date-time"
      required:
        - "id"
        - "content"
        - "createdBy"
        - "profileId"
        - "solvedAt"
        - "created_at"
        - "updated_at"
    GetByProfileCommentResponseDto:
      type:
        - "object"
      properties:
        comments:
          type:
            - "array"
          items:
            type:
              - "object"
            properties:
              id:
                type:
                  - "string"
                format: "uuid"
              content:
                type:
                  - "string"
                minLength: 1
              createdBy:
                type:
                  - "string"
                format: "uuid"
              profileId:
                type:
                  - "string"
                format: "uuid"
              isSolvable:
                default: false
                type:
                  - "boolean"
              isSolved:
                default: false
                type:
                  - "boolean"
              solvedAt:
                type:
                  - "string"
                  - "null"
                format: "date-time"
              solvedBy:
                type:
                  - "string"
                format: "uuid"
              created_at:
                type:
                  - "string"
                format: "date-time"
              updated_at:
                type:
                  - "string"
                format: "date-time"
              account:
                type:
                  - "object"
                properties:
                  username:
                    type:
                      - "string"
                required:
                  - "username"
            required:
              - "id"
              - "content"
              - "createdBy"
              - "profileId"
              - "solvedAt"
              - "created_at"
              - "updated_at"
              - "account"
      required:
        - "comments"
    ToggleSolveCommentResponseDto:
      type:
        - "object"
      properties:
        id:
          type:
            - "string"
          format: "uuid"
        content:
          type:
            - "string"
          minLength: 1
        createdBy:
          type:
            - "string"
          format: "uuid"
        profileId:
          type:
            - "string"
          format: "uuid"
        isSolvable:
          default: false
          type:
            - "boolean"
        isSolved:
          default: false
          type:
            - "boolean"
        solvedAt:
          type:
            - "string"
            - "null"
          format: "date-time"
        solvedBy:
          type:
            - "string"
          format: "uuid"
        created_at:
          type:
            - "string"
          format: "date-time"
        updated_at:
          type:
            - "string"
          format: "date-time"
      required:
        - "id"
        - "content"
        - "createdBy"
        - "profileId"
        - "solvedAt"
        - "created_at"
        - "updated_at"
...
    CreateEventFolderDto:
      type:
        - "object"
      properties:
        name:
          type:
            - "string"
          minLength: 1
        color:
          type:
            - "string"
          minLength: 7
          maxLength: 7
      required:
        - "name"
        - "color"
    CreateEventFolderResponseDto:
      type:
        - "object"
      properties:
        id:
          type:
            - "string"
          format: "uuid"
        name:
          type:
            - "string"
          minLength: 1
        color:
          type:
            - "string"
          minLength: 7
          maxLength: 7
        created_at:
          type:
            - "string"
          format: "date-time"
        updated_at:
          type:
            - "string"
          format: "date-time"
      required:
        - "id"
        - "name"
        - "color"
        - "created_at"
        - "updated_at"
...
    GetByIdEventFolderResponseDto:
      type:
        - "object"
      properties:
        id:
          type:
            - "string"
          format: "uuid"
        name:
          type:
            - "string"
          minLength: 1
        color:
          type:
            - "string"
          minLength: 7
          maxLength: 7
        created_at:
          type:
            - "string"
          format: "date-time"
        updated_at:
          type:
            - "string"
          format: "date-time"
        events:
          type:
            - "array"
          items:
            type:
              - "object"
            properties:
              id:
                type:
                  - "string"
                format: "uuid"
              name:
                type:
                  - "string"
                minLength: 1
              date:
                type:
                  - "string"
                format: "date-time"
              location:
                type:
                  - "string"
                minLength: 1
              folderId:
                type:
                  - "string"
                  - "null"
                format: "uuid"
              tagAssistedId:
                type:
                  - "string"
                format: "uuid"
              tagConfirmedId:
                type:
                  - "string"
                format: "uuid"
              supraEventId:
                type:
                  - "string"
                  - "null"
                format: "uuid"
              created_at:
                type:
                  - "string"
                format: "date-time"
              updated_at:
                type:
                  - "string"
                format: "date-time"
            required:
              - "id"
              - "name"
              - "date"
              - "location"
              - "folderId"
              - "tagAssistedId"
              - "tagConfirmedId"
              - "supraEventId"
              - "created_at"
              - "updated_at"
      required:
        - "id"
        - "name"
        - "color"
        - "created_at"
        - "updated_at"
        - "events"
    UpdateEventFolderDto:
      type:
        - "object"
      properties:
        name:
          type:
            - "string"
          minLength: 1
        color:
          type:
            - "string"
          minLength: 7
          maxLength: 7
      required:
        - "name"
        - "color"
    UpdateEventFolderResponseDto:
      type:
        - "object"
      properties:
        id:
          type:
            - "string"
          format: "uuid"
        name:
          type:
            - "string"
          minLength: 1
        color:
          type:
            - "string"
          minLength: 7
          maxLength: 7
        created_at:
          type:
            - "string"
          format: "date-time"
        updated_at:
          type:
            - "string"
          format: "date-time"
      required:
        - "id"
        - "name"
        - "color"
        - "created_at"
        - "updated_at"
    DeleteEventFolderResponseDto:
      type:
        - "object"
      properties:
        id:
          type:
            - "string"
          format: "uuid"
        name:
          type:
            - "string"
          minLength: 1
        color:
          type:
            - "string"
          minLength: 7
          maxLength: 7
        created_at:
          type:
            - "string"
          format: "date-time"
        updated_at:
          type:
            - "string"
          format: "date-time"
      required:
        - "id"
        - "name"
        - "color"
        - "created_at"
        - "updated_at"
    CreateEventDto:
      type:
        - "object"
      properties:
        name:
          type:
            - "string"
          minLength: 1
        date:
          type:
            - "string"
          format: "date-time"
        location:
          type:
            - "string"
          minLength: 1
        folderId:
          type:
            - "string"
            - "null"
          format: "uuid"
        subEvents:
          type:
            - "array"
          items:
            type:
              - "object"
            properties:
              name:
                type:
                  - "string"
                minLength: 1
              date:
                type:
                  - "string"
                format: "date-time"
              location:
                type:
                  - "string"
                minLength: 1
            required:
              - "name"
              - "date"
              - "location"
      required:
        - "name"
        - "date"
        - "location"
        - "folderId"
    CreateEventResponseDto:
      type:
        - "object"
      properties:
        id:
          type:
            - "string"
          format: "uuid"
        name:
          type:
            - "string"
          minLength: 1
        date:
          type:
            - "string"
          format: "date-time"
        location:
          type:
            - "string"
          minLength: 1
        folderId:
          type:
            - "string"
            - "null"
          format: "uuid"
        tagAssistedId:
          type:
            - "string"
          format: "uuid"
        tagConfirmedId:
          type:
            - "string"
          format: "uuid"
        supraEventId:
          type:
            - "string"
            - "null"
          format: "uuid"
        created_at:
          type:
            - "string"
          format: "date-time"
        updated_at:
          type:
            - "string"
          format: "date-time"
      required:
        - "id"
        - "name"
        - "date"
        - "location"
        - "folderId"
        - "tagAssistedId"
        - "tagConfirmedId"
        - "supraEventId"
        - "created_at"
        - "updated_at"
...
    GetByIdEventResponseDto:
      type:
        - "object"
      properties:
        id:
          type:
            - "string"
          format: "uuid"
        name:
          type:
            - "string"
          minLength: 1
        date:
          type:
            - "string"
          format: "date-time"
        location:
          type:
            - "string"
          minLength: 1
        folderId:
          type:
            - "string"
            - "null"
          format: "uuid"
        tagAssistedId:
          type:
            - "string"
          format: "uuid"
        tagConfirmedId:
          type:
            - "string"
          format: "uuid"
        supraEventId:
          type:
            - "string"
            - "null"
          format: "uuid"
        created_at:
          type:
            - "string"
          format: "date-time"
        updated_at:
          type:
            - "string"
          format: "date-time"
        subEvents:
          type:
            - "array"
          items:
            type:
              - "object"
            properties:
              id:
                type:
                  - "string"
                format: "uuid"
              name:
                type:
                  - "string"
                minLength: 1
              date:
                type:
                  - "string"
                format: "date-time"
              location:
                type:
                  - "string"
                minLength: 1
              folderId:
                type:
                  - "string"
                  - "null"
                format: "uuid"
              tagAssistedId:
                type:
                  - "string"
                format: "uuid"
              tagConfirmedId:
                type:
                  - "string"
                format: "uuid"
              supraEventId:
                type:
                  - "string"
                  - "null"
                format: "uuid"
              created_at:
                type:
                  - "string"
                format: "date-time"
              updated_at:
                type:
                  - "string"
                format: "date-time"
            required:
              - "id"
              - "name"
              - "date"
              - "location"
              - "folderId"
              - "tagAssistedId"
              - "tagConfirmedId"
              - "supraEventId"
              - "created_at"
              - "updated_at"
        supraEvent:
          type:
            - "object"
            - "null"
          properties:
            id:
              type:
                - "string"
              format: "uuid"
            name:
              type:
                - "string"
              minLength: 1
            date:
              type:
                - "string"
              format: "date-time"
            location:
              type:
                - "string"
              minLength: 1
            folderId:
              type:
                - "string"
                - "null"
              format: "uuid"
            tagAssistedId:
              type:
                - "string"
              format: "uuid"
            tagConfirmedId:
              type:
                - "string"
              format: "uuid"
            supraEventId:
              type:
                - "string"
                - "null"
              format: "uuid"
            created_at:
              type:
                - "string"
              format: "date-time"
            updated_at:
              type:
                - "string"
              format: "date-time"
          required:
            - "id"
            - "name"
            - "date"
            - "location"
            - "folderId"
            - "tagAssistedId"
            - "tagConfirmedId"
            - "supraEventId"
            - "created_at"
            - "updated_at"
      required:
        - "id"
        - "name"
        - "date"
        - "location"
        - "folderId"
        - "tagAssistedId"
        - "tagConfirmedId"
        - "supraEventId"
        - "created_at"
        - "updated_at"
        - "subEvents"
        - "supraEvent"
    UpdateEventDto:
      type:
        - "object"
      properties:
        name:
          type:
            - "string"
          minLength: 1
        folderId:
          type:
            - "string"
            - "null"
          format: "uuid"
        date:
          type:
            - "string"
          format: "date-time"
        location:
          type:
            - "string"
          minLength: 1
        subEvents:
          type:
            - "array"
          items:
            type:
              - "object"
            properties:
              name:
                type:
                  - "string"
                minLength: 1
              location:
                type:
                  - "string"
                minLength: 1
              date:
                type:
                  - "string"
                format: "date-time"
              id:
                oneOf:
                  - type:
                      - "string"
                    format: "uuid"
                  - type:
                      - "string"
                    enum:
                      - ""
            required:
              - "name"
              - "location"
              - "date"
              - "id"
      required:
        - "name"
        - "folderId"
        - "date"
        - "location"
        - "subEvents"
...
    DeleteEventResponseDto:
      type:
        - "object"
      properties:
        id:
          type:
            - "string"
          format: "uuid"
        name:
          type:
            - "string"
          minLength: 1
        date:
          type:
            - "string"
          format: "date-time"
        location:
          type:
            - "string"
          minLength: 1
        folderId:
          type:
            - "string"
            - "null"
          format: "uuid"
        tagAssistedId:
          type:
            - "string"
          format: "uuid"
        tagConfirmedId:
          type:
            - "string"
          format: "uuid"
        supraEventId:
          type:
            - "string"
            - "null"
          format: "uuid"
        created_at:
          type:
            - "string"
          format: "date-time"
        updated_at:
          type:
            - "string"
          format: "date-time"
      required:
        - "id"
        - "name"
        - "date"
        - "location"
        - "folderId"
        - "tagAssistedId"
        - "tagConfirmedId"
        - "supraEventId"
        - "created_at"
        - "updated_at"
...